+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Fix #5368: added support for additional ListOptions fields
 * Fix #5377: added a createOr and unlock function to provide a straight-forward replacement for createOrReplace.
 * Fix #5388: [crd-generator] Generate deterministic CRDs
+* Fix #5257: Add ErrorStreamMessage and StatusStreamMessage to ease mocking of pods/exec requests
 
 #### Dependency Upgrade
 * Fix #5373: Gradle base API based on v8.2.1